Refactor App component structure
