+++ src/App/App.fixture.tsx
@@ -1,4 +1,4 @@
 import React from "react";
-import App from "./App";
+import App from "./index";
 
 export default <App />;
